@@ -4,13 +4,14 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^3.9.3",
+    "@material-ui/core": "^4.3.2",
+    "@material-ui/styles": "^4.3.0",
     "lodash.throttle": "^4.1.1",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-pdf": "^4.0.5",
+    "react": "^16.9.0",
+    "react-dom": "^16.9.0",
+    "react-pdf": "^4.1.0",
     "react-scripts": "2.1.8",
-    "react-toastify": "^5.0.0-rc.5"
+    "react-toastify": "^5.3.2"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -30,6 +31,6 @@
     "not op_mini all"
   ],
   "devDependencies": {
-    "gh-pages": "^2.0.1"
+    "gh-pages": "^2.1.1"
   }
 }
